requirements.txt
addressable==1.4.2
awesome-slugify==1.6.5
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.4.1
bleach==1.4.3
blinker==1.3
cached-property==1.3.0
chardet==2.3.0
click==6.3
cmsplugin-filer==1.1.1
cycler==0.10.0
decorator==4.0.10
dj-database-url==0.4.1
Django==1.8.13
django-appconf==1.0.2
django-classy-tags==0.7.2
django-cms==3.2.5
django-contrib-comments==1.7.1
django-filer==1.2.2
django-formtools==1.0
django-jsonfield==1.0.0
django-mptt==0.8.4
django-polymorphic==0.8.1
django-reversion==1.8.7
django-sekizai==0.9.0
Django-Select2==4.3.2
django-treebeard==4.0.1
djangocms-admin-style==1.2.3
djangocms-attributes-field==0.0.2
djangocms-column==1.6.0
djangocms-file==1.0
djangocms-googlemap==0.5.1
djangocms-inherit==0.2.1
djangocms-installer==0.8.10
djangocms-link==1.8.0
djangocms-picture==0.2.0
djangocms-style==1.7.0
djangocms-teaser==0.2.0
djangocms-text-ckeditor==2.9.3
djangocms-video==1.0.0
easy-thumbnails==2.3
entrypoints==0.2.2
facebook-sdk==1.0.0
facebookinsights==0.3.4
facepy==1.0.8
filebrowser-safe==0.4.3
Flask==0.10.1
Flask-Admin==1.4.2
Flask-BabelEx==0.9.3
Flask-Cache==0.13.1
Flask-Gravatar==0.4.2
Flask-Login==0.2.9
Flask-Mail==0.9.1
Flask-Mistune==0.1.1
Flask-OAuthlib==0.9.2
Flask-Principal==0.4.0
Flask-Security==1.7.4
Flask-Stormpath==0.4.5
Flask-WTF==0.12
foreman==0.9.7
future==0.15.2
grappelli-safe==0.4.2
gunicorn==19.6.0
heroku==0.1.4
html5lib==0.9999999
httplib2==0.9.2
ipykernel==4.3.1
ipython==4.2.0
ipython-genutils==0.1.0
ipywidgets==5.1.5
isodate==0.5.4
itsdangerous==0.24
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.3.0
jupyter-console==4.1.1
jupyter-core==4.1.0
keyring==9.0
MarkupSafe==0.23
matplotlib==1.5.1
MechanicalSoup==0.4.0
mechanize==0.2.5
Mezzanine==4.1.0
mistune==0.7.3
mongoengine==0.10.1
nbconvert==4.2.0
nbformat==4.0.1
networkx==1.11
notebook==4.2.1
numpy==1.11.1
oauth2==1.9.0.post1
oauth2client==1.2
oauthlib==1.1.2
passlib==1.6.5
pickleshare==0.7.2
Pillow==3.1.1
plotly==1.12.2
PyDispatcher==2.0.5
pygal==2.2.3
Pygments==2.1.3
PyJWT==1.4.0
pylev==1.3.0
pymongo==2.9.2
pyparsing==2.1.5
pyreadline==2.1
PyRSS2Gen==1.1
pyshorteners==0.5.8
python-dateutil==1.5
pytz==2016.4
pywin32-ctypes==0.0.1
pyzmq==15.2.0
qtconsole==4.2.1
quokka-flask-admin==1.3.0
quokka-flask-htmlbuilder==0.10
quokka-flask-mongoengine==0.7.4
quokka-speaklater==1.3.1
Quokka-Themes==0.1.4
rauth==0.7.2
regex==2016.6.24
requests==2.9.1
requests-futures==0.9.7
requests-oauthlib==0.6.1
robobrowser==0.5.3
runipy==0.1.5
selenium==2.53.6
shiftpy==0.1.3
simplegeneric==0.8.1
simplejson==3.8.2
six==1.10.0
speaklater==1.3
stormpath==2.1.1
tornado==4.3
traitlets==4.2.1
tzlocal==1.2.2
Unidecode==0.4.19
virtualenv==15.0.2
vpnotebook==0.1.3
vpython==0.3.11
Werkzeug==0.11.10
widgetsnbextension==1.2.3
WTForms==2.1
